requirement.txt
asn1crypto==0.24.0
attrs==18.1.0
autobahn==18.6.1
Automat==0.7.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.2.2
Flask==1.0.2
future==0.16.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pkg-resources==0.0.0
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycparser==2.18
pyOpenSSL==18.0.0
python-dateutil==2.7.3
python-telegram-bot==10.1.0
requests==2.19.1
service-identity==17.0.0
six==1.11.0
Twisted==18.4.0
txaio==2.10.0
urllib3==1.23
watson-developer-cloud==1.5.0
Werkzeug==0.14.1
zope.interface==4.5.0
